+++ rust/src/api/keystone.rs
@@ -35,10 +35,10 @@ pub fn encode_pczt_ur_parts(
 /// scan ended (cancel, back button, mid-stream error).
 ///
 /// Marked `#[frb(sync)]` so the Dart caller does not race with the camera:
-/// `_AnimatedUrScanScreenState.initState` needs the Rust `UR_SESSION` to be
-/// clean **before** the first `onDetect` callback fires, and a fire-and-forget
-/// `Future` provides no such ordering guarantee. The Rust body is a single
-/// mutex lock + `None` assignment, so it's trivially non-blocking.
+/// QR scan screen entry needs the Rust `UR_SESSION` to be clean **before** the
+/// first `onDetect` callback fires, and a fire-and-forget `Future` provides no
+/// such ordering guarantee. The Rust body is a single mutex lock + `None`
+/// assignment, so it's trivially non-blocking.
 #[flutter_rust_bridge::frb(sync)]
 pub fn reset_ur_session() {
     keystone::reset_ur_session();
